requirements.txt
fastapi[standard]
uvicorn
pydantic[email]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
httpx
python-jose[cryptography]
python-dotenv
gunicorn
pytest
